Refactor operators module
Create common base class to serve as a wrapper around rust-implemented
operators. This wrapper allows us to inspect the types passed, and
unwrap them if we can to improve performance.

We move the metric protocols and typevar defitions into the operators
module in order to make them accessible as actual symbols, rather than
defining them in a type stub where they cannot be imported.